refactoring Auth, Main, Camera
