+++ package.json
@@ -15,7 +15,6 @@
     "@types/react-dom": "^18.0.11",
     "bootstrap": "^5.3.0",
     "firebase": "^9.20.0",
-    "node-sass": "^7.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
@@ -52,6 +51,7 @@
     "@typescript-eslint/eslint-plugin": "^5.59.2",
     "@typescript-eslint/parser": "^5.59.2",
     "eslint": "^8.39.0",
-    "eslint-plugin-react": "^7.32.2"
+    "eslint-plugin-react": "^7.32.2",
+    "node-sass": "npm:sass@^1.69.5"
   }
 }
