+++ src/App.tsx
@@ -2,17 +2,17 @@ import "./common/styles/App.scss";
 
 import React from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-import { SignIn } from "./auth/components/SignIn";
-import { SignUp } from "./auth/components/SignUp";
-import { Main } from "./common/components/Main";
-import { CreateCamera } from "./camera/components/CreateCamera";
-import { Viewer } from "./viewer/components/Viewer";
+import { SignIn } from "./auth/view/SignIn";
+import { SignUp } from "./auth/view/SignUp";
+import { Main } from "./common/view/Main";
+import { CreateCamera } from "./camera/view/CreateCamera";
+import { Viewer } from "./viewer/view/Viewer";
 import { AppLayout } from "./common/layouts/AppLayout";
-import { Camera } from "./camera/components/Camera";
+import { Camera } from "./camera/view/Camera";
 import { ConnectionProvider } from "./common/contexts/ConnectionContext";
 import { useAuthContext } from "./common/contexts/AuthContext";
-import { CreateViewer } from "./viewer/components/CreateViewer";
-import { History } from "./history/components/History";
+import { CreateViewer } from "./viewer/view/CreateViewer";
+import { History } from "./history/view/History";
 
 function App() {
   const {user} = useAuthContext();
