+++ src/auth/view/SignIn.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link, useNavigate } from "react-router-dom";
 
-import { AuthForm } from "./AuthForm";
+import { AuthForm } from "../components/AuthForm";
 
 import { signInEmail } from "../../common/functions/auth";
 
@@ -19,16 +19,11 @@ export function SignIn () {
 
   return <div className="auth body-content container-fluid pt-3">
     <div className="card">
-
-        <h1 className="card-header">Sign-in with Your Account</h1>
-
+      <div className="card-header fs-5">Sign-in with Your Account</div>
       <div className="card-body">
       <AuthForm buttonText="Sign In" onSubmit={onSubmit} />
-      
       </div>
     </div>
-    
-
     <Link to="/sign-up">
       <button type="button" className="btn btn-link mx-auto d-block">
         or click here to sign up for new account
