+++ src/common/layouts/AppLayout.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import {Outlet} from "react-router-dom";
 
-import {Header} from "../components/Header";
-import {Sidebar} from "../components/Sidebar";
+import {Header} from "./Header";
+import {Sidebar} from "./Sidebar";
 
 
 export function AppLayout () {
