+++ src/common/layouts/Sidebar.tsx
@@ -38,7 +38,6 @@ export function Sidebar () {
     aria-labelledby="offcanvasNavbarLabel"
   >
     <div className="offcanvas-header justify-content-end">
-      {/* <h5 className="offcanvas-title" id="offcanvasNavbarLabel">Menu</h5> */}
       <button type="button" className="btn-close" data-bs-dismiss="offcanvas" aria-label="Close"></button>
     </div>
     <div className="offcanvas-body">
