+++ src/common/styles/App.scss
@@ -57,55 +57,55 @@ a {
 
 
 
-h1, .h1 {
-  font-size: calc(1.125rem + 1.5vw);
-  font-weight: 400;
-}
-@media (min-width: 1200px) {
-  h1, .h1 {
-    font-size: 1rem;
-  }
-}
+// h1, .h1 {
+//   font-size: calc(1.125rem + 1.5vw);
+//   font-weight: 400;
+// }
+// @media (min-width: 1200px) {
+//   h1, .h1 {
+//     font-size: 2.5rem;
+//   }
+// }
 
-h2, .h2 {
-  font-size: calc(1.1rem + 0.9vw);
-  font-weight: 400;
-}
-@media (min-width: 1200px) {
-  h2, .h2 {
-    font-size: 2rem;
-  }
-}
+// h2, .h2 {
+//   font-size: calc(1.1rem + 0.9vw);
+//   font-weight: 400;
+// }
+// @media (min-width: 1200px) {
+//   h2, .h2 {
+//     font-size: 2rem;
+//   }
+// }
 
-h3, .h3 {
-  font-size: calc(1.075rem + 0.6vw);
-  font-weight: 400;
-}
-@media (min-width: 1200px) {
-  h3, .h3 {
-    font-size: 1.75rem;
-  }
-}
+// h3, .h3 {
+//   font-size: calc(1.075rem + 0.6vw);
+//   font-weight: 400;
+// }
+// @media (min-width: 1200px) {
+//   h3, .h3 {
+//     font-size: 1.75rem;
+//   }
+// }
 
-h4, .h4 {
-  font-size: calc(1.05rem + 0.3vw);
-  font-weight: 400;
-}
-@media (min-width: 1200px) {
-  h4, .h4 {
-    font-size: 1.5rem;
-  }
-}
+// h4, .h4 {
+//   font-size: calc(1.05rem + 0.3vw);
+//   font-weight: 400;
+// }
+// @media (min-width: 1200px) {
+//   h4, .h4 {
+//     font-size: 1.5rem;
+//   }
+// }
 
-h5, .h5 {
-  font-size: 1rem;
-  font-weight: 400;
-}
+// h5, .h5 {
+//   font-size: 1rem;
+//   font-weight: 400;
+// }
 
-h6, .h6 {
-  font-size: 0.75rem;
-  font-weight: 400;
-}
+// h6, .h6 {
+//   font-size: 0.75rem;
+//   font-weight: 400;
+// }
 
 
 .navbar {
