+++ src/index.tsx
@@ -11,8 +11,7 @@ const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
-  
-    <AuthProvider>
-      <App />
-    </AuthProvider>
+  <AuthProvider>
+    <App />
+  </AuthProvider>
 );
